Небольшие исправления

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -33,5 +33,5 @@ public interface UserStorage {
 
     boolean removeRequest(Object... params);
 
-    List<Long> getBestRepetitionUserIds(int userId);
+    List<Long> getBestRepetitionUserIds(Long userId);
 }
